create abstract classes for device and model manages, and use them as components rather than direct import
