+++ nvflare/edge/assessors/model_manager.py
@@ -12,188 +12,75 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import time
-from typing import Dict, Optional
+from abc import ABC, abstractmethod
+from typing import Any
 
-import numpy as np
-
-from nvflare.apis.dxo import DXO, DataKind
 from nvflare.apis.fl_component import FLComponent
 from nvflare.apis.fl_context import FLContext
-from nvflare.app_common.abstract.model import make_model_learnable
-from nvflare.app_common.app_constant import AppConstants
-from nvflare.app_common.app_event_type import AppEventType
-from nvflare.edge.aggregators.model_update_dxo import ModelUpdateDXOAggregator
-from nvflare.edge.mud import ModelUpdate
 
 
-class _ModelState:
-    def __init__(self, aggr: ModelUpdateDXOAggregator):
-        self.aggregator = aggr
-        self.devices = {}
-        self.last_update_time = None
+class ModelManager(FLComponent, ABC):
+    """Abstract base class for model managers in federated learning.
 
-    def accept(self, model_update: ModelUpdate, fl_ctx: FLContext):
-        self.last_update_time = time.time()
-        self.devices.update(model_update.devices)
-        return self.aggregator.accept(model_update.update, fl_ctx)
+    This class defines the interface that all model managers must implement.
+    Model managers are responsible for handling model updates, aggregation,
+    and version control in federated learning workflows.
+    """
 
-
-class ModelManager(FLComponent):
-    def __init__(
-        self,
-        num_updates_for_model: int,
-        max_model_version: int,
-        max_model_history: int,
-        global_lr: float,
-        staleness_weight: bool,
-    ):
-        """Initialize the ModelManager.
-        The aggregation scheme and weights are calculated following FedBuff paper "Federated Learning with Buffered Asynchronous Aggregation".
-        The staleness_weight can be enabled to apply staleness weighting to model updates.
-        Args:
-            num_updates_for_model (int): Number of updates required before generating a new model version.
-            max_model_version (int): Maximum number of model versions allowed in the workflow.
-            max_model_history (int): Maximum number of historical model versions to keep in memory.
-            global_lr (float): Global learning rate for model aggregation.
-            staleness_weight (bool): Whether to apply staleness weighting to model updates.
-        """
-
+    def __init__(self):
         FLComponent.__init__(self)
+        """Initialize the ModelManager.
+        ModelManager keeps track of two things: 
+        - current_model holding the current global model
+        - updates containing all received updates for updating the global model 
+        """
         self.current_model = None
-        self.current_model_version = 0
-        self.updates = {}  # model_version => _ModelState
-        self.num_updates_for_model = num_updates_for_model
-        self.num_updates_counter = 0
-        self.max_model_version = max_model_version
-        self.max_model_history = max_model_history
-        self.global_lr = global_lr
-        self.staleness_weight = staleness_weight
+        self.updates = {}
 
-    def initialize_model(self, model: DXO):
-        self.current_model = model
-        self.updates[self.current_model_version] = _ModelState(ModelUpdateDXOAggregator())
+    @abstractmethod
+    def initialize_model(self, model: Any, fl_ctx: FLContext) -> None:
+        """Initialize the model manager with an initial model.
 
+        Args:
+            model: The initial model
+            fl_ctx: FLContext object
+
+            Returns: none
+        """
+        pass
+
+    @abstractmethod
     def generate_new_model(self, fl_ctx: FLContext) -> None:
-        # New model generated based on the current global weights and all updates
-        new_model = {}
-        self.current_model_version += 1
-        old_model_versions = []
+        """Generate a new model version based on accumulated updates.
 
-        # counter to confirm the number of updates
-        num_updates = 0
+        Args:
+            fl_ctx: FLContext object
 
-        if self.current_model_version == 1:
-            # Initial global weights
-            new_model = self.current_model.data
-        else:
-            # Aggregate all updates
-            for v, ms in self.updates.items():
-                if self.staleness_weight:
-                    weight = 1 / (1 + (self.current_model_version - v) ** 0.5)
-                else:
-                    weight = 1.0
-                aggr = ms.aggregator
-                # Add the dict to new_model by multiplying the weight and dividing by the count
-                update_dict = aggr.dict
-                count = aggr.count
+            Returns: none
+        """
+        pass
 
-                if count > 0:
-                    # aggregate updates
-                    for key, value in update_dict.items():
-                        # apply weight and divide by count
-                        value = weight * value / count
-                        # update the new model
-                        if key not in new_model:
-                            new_model[key] = value
-                        else:
-                            new_model[key] = new_model[key] + value
-                # If too old, remove it
-                if self.current_model_version - v >= self.max_model_history:
-                    old_model_versions.append(v)
+    @abstractmethod
+    def process_updates(self, model_updates: Any, fl_ctx: FLContext) -> bool:
+        """Process incoming model updates from clients.
 
-                # Reset aggr after counting its contribution
-                ms.aggregator.reset(fl_ctx)
-                num_updates += count
+        Args:
+            model_updates: updates collected from clients
+            fl_ctx: FLContext object
 
-            # Add the aggregated updates to the current global weights
-            global_weights = self.current_model.data
-            for key, value in new_model.items():
-                if key not in global_weights:
-                    self.log_error(fl_ctx, f"key {key} not in new model")
-                    continue
-                new_model[key] = np.array(global_weights[key]) + value * self.global_lr
+        Returns:
+            bool: Whether the updates were successfully processed
+        """
+        pass
 
-        # create the ModelState for the new model version
-        self.updates[self.current_model_version] = _ModelState(ModelUpdateDXOAggregator())
-        self.log_info(fl_ctx, f"generated new model version {self.current_model_version} with {num_updates} updates")
+    @abstractmethod
+    def get_current_model(self, fl_ctx: FLContext) -> Any:
+        """Get the current model.
 
-        if old_model_versions:
-            self.log_info(fl_ctx, f"removed old model versions {old_model_versions}")
+        Args:
+            fl_ctx: FLContext object
 
-        for v in old_model_versions:
-            self.updates.pop(v)
-
-        # update the current model
-        # convert new_model items from numpy arrays to lists for serialization
-        new_model = {k: v.tolist() if isinstance(v, np.ndarray) else v for k, v in new_model.items()}
-        self.current_model = DXO(data_kind=DataKind.WEIGHTS, data=new_model)
-
-        # reset the num_updates_counter
-        self.num_updates_counter = 0
-
-        # set fl_ctx and fire the event
-        # wrap new_model to a learnable
-        learnable = make_model_learnable(new_model, {})
-        fl_ctx.set_prop(AppConstants.GLOBAL_MODEL, learnable, private=True, sticky=True)
-        fl_ctx.set_prop(AppConstants.CURRENT_ROUND, self.current_model_version, private=True, sticky=True)
-        self.fire_event(AppEventType.GLOBAL_WEIGHTS_UPDATED, fl_ctx)
-
-    def process_updates(self, model_updates: Dict[int, ModelUpdate], fl_ctx: FLContext) -> bool:
-        accepted = True
-        for model_version, model_update in model_updates.items():
-            if model_version <= 0:
-                continue
-
-            if not model_update:
-                self.log_error(fl_ctx, f"bad child update version {model_version}: no update data")
-                continue
-
-            if self.current_model_version - model_version > self.max_model_history:
-                # this version is too old
-                self.log_info(
-                    fl_ctx,
-                    f"dropped child update version {model_version}. Current version {self.current_model_version}",
-                )
-                continue
-
-            model_state = self.updates.get(model_version)
-            if not model_state:
-                self.log_error(fl_ctx, f"No model state for version {model_version}")
-                continue
-
-            accepted = model_state.accept(model_update, fl_ctx)
-            self.log_info(
-                fl_ctx,
-                f"processed child update V{model_version} with {len(model_update.devices)} devices: {accepted=}",
-            )
-
-            # update the global num_updates_counter
-            self.num_updates_counter += len(model_update.devices)
-
-        current_model_state = self.updates.get(self.current_model_version)
-        if isinstance(current_model_state, _ModelState):
-            if self.num_updates_counter >= self.num_updates_for_model:
-                self.log_info(
-                    fl_ctx,
-                    f"Globally got {self.num_updates_counter} updates: generate new model version",
-                )
-                self.generate_new_model(fl_ctx)
-
-        return accepted
-
-    def should_generate_initial_model(self, num_devices: int) -> bool:
-        return self.current_model_version == 0 and num_devices >= self.device_selection_size
-
-    def get_current_model(self) -> Optional[DXO]:
+        Returns:
+            The current model
+        """
         return self.current_model
